Make the remote HEAD symbolic ref optional
### What
Make the remote HEAD symbolic ref optional and default the base branch
to master when it is not present.

### Why
In some cases the symbolic ref can be lost and the scripts print out
errors about that repeatedly. For most operations though it has no
impact. Defaulting to master which is the most common default branch of
git repos should catch some of the cases where it is needed and the ref
was lost.

It might be ideal to leave it as an empty string though in those cases.